// frontend/src/api/model/concourseItemStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Qualis API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Enum for concourse item curation status.
 */
export type ConcourseItemStatus = (typeof ConcourseItemStatus)[keyof typeof ConcourseItemStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ConcourseItemStatus = {
    proposed: 'proposed',
    accepted: 'accepted',
    rejected: 'rejected',
} as const;